chore(web): extract ImageEditor mask/boxes/color-grade tools into hooks (sc-9752)
Deferred remainder of sc-8854 (F-052 god components). Decompose the three
per-tool logic clusters in ImageEditor.jsx into dedicated hooks + cycle-free
pure-helper modules under src/screens/imageEditor/, following the sc-9750
hooks/ extraction style. Strictly behavior-preserving — no UX/render change.

Extracted:
- useColorGradeTool: owns colorAdjust/mode/levels/curves/channel state, the
  live-preview cache + histogram effects, and the Apply bake. Pure math moves
  to colorGradeMath.js (gradePixel/applyColorAdjustments/isIdentityAdjust/
  konvaColorFilter/COLOR_ADJUSTMENTS/IDENTITY_COLOR_ADJUST).
- useBoxesTool: owns boxes/selectedBoxId/boxColor/boxDraft + the Konva node /
  transformer refs + the boxId sequence, the draw/select/drag/transform
  handlers, and the transformer-binding effect. Pure box geometry/validation/
  adapters move to boxGeometry.js.
- useMaskTool: owns the brush-stroke + smart-select mask state, the brush/select
  pointer handlers, rasterize/refine ops, the SAM3 smart-select run, and the
  leave-paint-mode effect. Pure helpers move to maskShared.js
  (buildSegmentJobBody/rectToSegmentBox/tintMaskRgbaInPlace/maskHasContent).

Ref-mirror preservation (the crux): the boxes hook still owns boxesRef/
boxColorRef/boxIdRef synced by the same [boxes]/[boxColor] effects, and returns
them so captureSnapshot reads / applyHistoryAux writes the SAME refs as before;
the mask hook keeps maskPaintingRef + selectDrawingRef/selectStartRef as live
gesture latches. Late-defined shared callbacks (checkpoint / stagePointToImage /
replaceLayerImage / runAiOp) are passed via stable ref bridges so each hook
invokes the latest closure exactly as the inline code did, with a fixed hook
call order. Effect dependency arrays + eslint-disable comments are preserved
verbatim.

ImageEditor.jsx re-exports every extracted pure symbol so its public surface —
and ImageEditor.test.jsx's imports — stay byte-for-byte unchanged.

Tests: full web vitest suite 1178 -> 1185 (90 -> 91 files), all green; added
imageEditorModules.test.js pinning the re-export identity contract + standalone
behavior. lint 0 errors (46 pre-existing warnings, none new). Production build
clean. ImageEditor.jsx: 928 lines removed, 252 added.
